Trying to figure out where the threading lives.
I think I like it being totally external to the Console, so that the
tests are by default deterministic / single-threaded. But it's still a
little funny to have a callback for shutting down the main loop, and for
start to not block. Going to tug on this a little more, I think.