Undo Extractor mess, yay! 🌯
Setting up to honor file permissions in archive files.

We need the JDK's lib/jspawnhelper to be executable, otherwise
kotlinc-native blows up with a misleading ENOENT trying to run
/usr/bin/xcrun.

While we could just make it executable more manually, this feels like
something we should get right, so we don't have to think about it in the
future.